Update OOP.md
Based on the example in the main func, I think this was a typo and it meant to say SHADOWS because "derived.a" is shadowing "base.a", which is being shadowed. Hopefully this improves the readability because I had to re-read the example a few times to make sure I understood it.